refactor: move trimStringSettings and validateSettings functions to separate utility files for better organization
